+++ items/coin.gd
@@ -1,0 +1,6 @@
+# coin.gd
+extends "res://items/Item.gd"
+
+
+func _ready():
+	self.item_id = GameData.PICKUPS.COIN